Skip test_mulled_build on regular unit tests
And move commonly used skip_if_environ/external_dependency_management
to test/unit/tool_util/util.py
