+++ test/unit/tool_util/mulled/test_get_tests.py
@@ -1,6 +1,6 @@
 from galaxy.tool_util.deps.mulled.get_tests import deep_test_search, find_anaconda_versions, get_alternative_versions, get_anaconda_url, get_commands_from_yaml, get_run_test, get_test_from_anaconda, hashed_test_search, main_test_search, open_recipe_file, prepend_anaconda_url
 from galaxy.util import smart_str
-from ..test_conda_resolution import external_dependency_management
+from ..util import external_dependency_management
 
 TEST_RECIPE = r"""
 {% set name = "eagle" %}
