+++ test/unit/tool_util/mulled/test_mulled_build.py
@@ -1,10 +1,13 @@
 import pytest
+
 from galaxy.tool_util.deps.mulled.mulled_build import (
     any_target_requires_extended_base,
     build_target,
 )
+from ..util import external_dependency_management
 
 
+@external_dependency_management
 @pytest.mark.parametrize("target,requires_extended", [
     ('maker', True),
     ('samtools', False)
