+++ test/unit/tool_util/mulled/test_mulled_search.py
@@ -1,7 +1,7 @@
 import pytest
 
 from galaxy.tool_util.deps.mulled.mulled_search import CondaSearch, get_package_hash, GitHubSearch, QuaySearch, run_command, singularity_search
-from ..test_conda_resolution import external_dependency_management
+from ..util import external_dependency_management
 
 
 @external_dependency_management
