+++ test/unit/tool_util/mulled/test_mulled_update_singularity_containers.py
@@ -6,7 +6,7 @@ import pytest
 
 from galaxy.tool_util.deps.mulled.mulled_update_singularity_containers import docker_to_singularity, get_list_from_file, singularity_container_test
 from galaxy.util import which
-from ..test_conda_resolution import external_dependency_management
+from ..util import external_dependency_management
 
 
 @external_dependency_management
